+++ lib/tad/dhcp/tad_dhcp_impl.h
@@ -96,7 +96,8 @@ extern char* dhcp_get_param_cb (csap_p csap_descr, int level, const char *param)
  * @return 
  *      quantity of read octets, or -1 if error occured, 0 if timeout expired. 
  */ 
-extern int dhcp_read_cb (csap_p csap_descr, int timeout, char *buf, int buf_len);
+extern int dhcp_read_cb(csap_p csap_descr, int timeout, char *buf,
+                        size_t buf_len);
 
 /**
  * Callback for write data to media of DHCP CSAP. 
@@ -108,7 +109,7 @@ extern int dhcp_read_cb (csap_p csap_descr, int timeout, char *buf, int buf_len)
  * @return 
  *      quantity of written octets, or -1 if error occured. 
  */ 
-extern int dhcp_write_cb (csap_p csap_descr, char *buf, int buf_len);
+extern int dhcp_write_cb(csap_p csap_descr, char *buf, size_t buf_len);
 
 /**
  * Callback for write data to media of DHCP CSAP and read
@@ -125,8 +126,8 @@ extern int dhcp_write_cb (csap_p csap_descr, char *buf, int buf_len);
  *      quantity of read octets, or -1 if error occured, 0 if timeout expired. 
  */ 
 extern int dhcp_write_read_cb(csap_p csap_descr, int timeout,
-                              char *w_buf, int w_buf_len,
-                              char *r_buf, int r_buf_len);
+                              char *w_buf, size_t w_buf_len,
+                              char *r_buf, size_t r_buf_len);
 
 
 /**
