+++ lib/tad/file/tad_file_impl.h
@@ -80,7 +80,7 @@ extern char* file_get_param_cb(int csap_id, int level, const char *param);
  * @return 
  *      quantity of read octets, or -1 if error occured, 0 if timeout expired. 
  */ 
-extern int file_read_cb(csap_p csap_descr, int timeout, char *buf, int buf_len);
+extern int file_read_cb(csap_p csap_descr, int timeout, char *buf, size_t buf_len);
 
 /**
  * Callback for write data to media of 'file' CSAP. 
@@ -92,7 +92,7 @@ extern int file_read_cb(csap_p csap_descr, int timeout, char *buf, int buf_len);
  * @return 
  *      quantity of written octets, or -1 if error occured. 
  */ 
-extern int file_write_cb(csap_p csap_descr, char *buf, int buf_len);
+extern int file_write_cb(csap_p csap_descr, char *buf, size_t buf_len);
 
 /**
  * Callback for write data to media of 'file' CSAP and read
@@ -109,8 +109,8 @@ extern int file_write_cb(csap_p csap_descr, char *buf, int buf_len);
  *      quantity of read octets, or -1 if error occured, 0 if timeout expired. 
  */ 
 extern int file_write_read_cb(csap_p csap_descr, int timeout,
-                              char *w_buf, int w_buf_len,
-                              char *r_buf, int r_buf_len);
+                              char *w_buf, size_t w_buf_len,
+                              char *r_buf, size_t r_buf_len);
 
 
 /**
