+++ lib/tad/snmp/tad_snmp_impl.h
@@ -137,7 +137,7 @@ extern char* snmp_get_param_cb (int csap_id, int level, const char *param);
  * @return 
  *      quantity of read octets, or -1 if error occured, 0 if timeout expired. 
  */ 
-extern int snmp_read_cb (csap_p csap_descr, int timeout, char *buf, int buf_len);
+extern int snmp_read_cb (csap_p csap_descr, int timeout, char *buf, size_t buf_len);
 
 /**
  * Callback for write data to media of 'snmp' CSAP. 
@@ -149,7 +149,7 @@ extern int snmp_read_cb (csap_p csap_descr, int timeout, char *buf, int buf_len)
  * @return 
  *      quantity of written octets, or -1 if error occured. 
  */ 
-extern int snmp_write_cb (csap_p csap_descr, char *buf, int buf_len);
+extern int snmp_write_cb (csap_p csap_descr, char *buf, size_t buf_len);
 
 /**
  * Callback for write data to media of 'snmp' CSAP and read
@@ -166,8 +166,8 @@ extern int snmp_write_cb (csap_p csap_descr, char *buf, int buf_len);
  *      quantity of read octets, or -1 if error occured, 0 if timeout expired. 
  */ 
 extern int snmp_write_read_cb (csap_p csap_descr, int timeout,
-                     char *w_buf, int w_buf_len,
-                     char *r_buf, int r_buf_len);
+                     char *w_buf, size_t w_buf_len,
+                     char *r_buf, size_t r_buf_len);
 
 
 /**
